agriassist/requirements.txt
# Core Framework
setuptools
wheel
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0

google-generativeai==0.8.6
langchain-core==0.3.39
pillow>=10.0.0
matplotlib==3.10.8


# Geospatial
shapely==2.0.2
pyproj==3.6.1
numpy<2.0.0

# Templates
jinja2==3.1.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
pydantic>=2.7.4
pydantic-settings==2.1.0
email-validator>=2.1.0

# Graph Database
neo4j==5.16.0

# Machine Learning & Computer Vision (COMMEMTED OUT for Render 512MB RAM)
# To use these, you need at least 2GB RAM (Railway Pro or Hugging Face)
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch
# torchvision
# ultralytics
# opencv-python-headless

# Communication
twilio==9.0.0

# Database migration
psycopg2-binary